+++ src/components/Content/FooterContent.vue
@@ -1,5 +1,5 @@
 <template>
-    <footer class="uf-main uk-section uk-section-small uk-section-muted uk-text-center">
+    <footer class="uk-section uk-section-small uk-section-muted uk-text-center">
         <hr />
         <p class="uk-text-small uk-text-center" data-test="slot">
             <slot></slot>