+++ src/components/Content/MainContent.vue
@@ -1,9 +1,9 @@
 <template>
-    <div
-        class="uf-main uk-section uk-section-default uk-section-muted"
+    <main
+        class="uk-section uk-section-default uk-section-muted"
         uk-height-viewport="expand: true">
         <div class="uk-container" data-test="slot">
             <slot></slot>
         </div>
-    </div>
+    </main>
 </template>